Refactor project items
- Move item_type to ProjectItem base class from the concere project
items. All of them should have this attribute and this is the way to
enforce it.
- Add notify_destination() method to ProjectItem. This moves the
responsibility of informing user of item connections from
DesignQGraphicsView to the items themselves leading to less coupling,
clear interfaces and less spaghetti code.
- Added unit tests for item_type and notify_destination(). This means
that we now have at least two tests for each concrete project item.

Re #413

+++ spinetoolbox/graphics_items.py
@@ -240,10 +240,10 @@ class ProjectItemIcon(QGraphicsRectItem):
 
         Args:
             toolbox (ToolBoxUI): QMainWindow instance
-            x (int): Icon x coordinate
-            y (int): Icon y coordinate
-            w (int): Icon width
-            h (int): Icon height
+            x (float): Icon x coordinate
+            y (float): Icon y coordinate
+            w (float): Icon width
+            h (float): Icon height
             name (str): Item name
         """
         super().__init__()
